ci: (Maybe) fix executable uploads
The `actions/upload-artifact@v2` action doesn't seem to support EXTGLOB
patterns, even though it uses the `minimatch` library under the hood, which has
them enabled by default.

My solution is to distribute the *.d files alongside the executables, which the
previous solution would've avoided (had it worked). This is preferable to
spending more time trying to make it work for not much gain.